Fix CUDA compatibility by defaulting to float16 dtype

+++ mira/streaming_model.py
@@ -8,11 +8,11 @@ from mira.utils import clear_cache
 class MiraTTSStreaming:
     """MiraTTS with real chunked streaming via LMDeploy stream_infer"""
 
-    def __init__(self, model_dir="YatharthS/MiraTTS", tp=1, enable_prefix_caching=True, cache_max_entry_count=0.2):
+    def __init__(self, model_dir="YatharthS/MiraTTS", tp=1, enable_prefix_caching=True, cache_max_entry_count=0.2, dtype='float16'):
         backend_config = TurbomindEngineConfig(
             cache_max_entry_count=cache_max_entry_count,
             tp=tp,
-            dtype='bfloat16',
+            dtype=dtype,
             enable_prefix_caching=enable_prefix_caching
         )
         self.pipe = pipeline(model_dir, backend_config=backend_config)
